project startup

+++ modules/db.js
@@ -8,28 +8,50 @@ const pool = new pg.Pool({
  });
 
 // database methods -------------------------
-let dbMethods = {}; //create empty object
+let dbMethods = {};
 
-// ------------------------------------
+// ----------- BLOGPOSTS -------------------------
 dbMethods.getAllBlogPosts = function() {
     let sql = "SELECT * FROM blogposts";	
-	return pool.query(sql); //return the promise	
+	return pool.query(sql);	
 }
 
-// ------------------------------------
 dbMethods.createBlogPost = function(heading, blogtext, userid) {  
     let sql = "INSERT INTO blogposts (id, date, heading, blogtext, userid) VALUES(DEFAULT, DEFAULT, $1, $2, $3) returning *";
 	let values = [heading, blogtext, userid];	
-    return pool.query(sql, values); //return the promise
+    return pool.query(sql, values); 
 }
 
-// ------------------------------------
-dbMethods.deleteBlogPost = function(id) {  
-    let sql = "DELETE FROM blogposts WHERE id = $1 RETURNING *";
+dbMethods.deleteBlogPost = function(id, userid) {  
+    let sql = "DELETE FROM blogposts WHERE id = $1 AND userid = $2 RETURNING *";
+	let values = [id, userid];	
+    return pool.query(sql, values); 
+}
+
+// --------------- USERS ---------------------
+dbMethods.getAllUsers = function() {
+    let sql = "SELECT id, username FROM users";	
+	return pool.query(sql); 	
+}
+
+dbMethods.getUser = function(username) {
+    let sql = "SELECT * FROM users WHERE username = $1";
+    let values = [username];	
+	return pool.query(sql, values);	
+}
+
+dbMethods.createUser = function(username, password, salt) {  
+    let sql = "INSERT INTO users (id, username, password, salt) VALUES(DEFAULT, $1, $2, $3) returning *";
+	let values = [username, password, salt];	
+    return pool.query(sql, values); 
+}
+
+dbMethods.deleteUser = function(id) {  
+    let sql = "DELETE FROM users WHERE id = $1 RETURNING *";
 	let values = [id];	
-    return pool.query(sql, values); //return the promise
+    return pool.query(sql, values); 
 }
 
-// export dbMethods -------------------------
+// -------- EXPORTS -----------------
 module.exports = dbMethods;
